fix(analytics): move slugify off the sanity-importing module
The previous round imported `slugify` from `lib/seo/legacy-redirect.ts` into
`<EmptyStateUndoAnalytics>`, a client component. That module imports
`fetchGroq` -> `lib/sanity/client`, which constructs a Sanity client at module
scope, so every story rendering the component threw
"Configuration must contain `projectId`" in Storybook, where no
NEXT_PUBLIC_SANITY_PROJECT_ID exists.

The visual-regression suite failed on 14 tests across calendarwidget,
eventsbrowser and hulpfinder with zero snapshot diffs (2826 of 2826 passed) --
a JS exception at story render, not a pixel change.

`slugify` now lives in `lib/utils/slugify.ts`, dependency-free on purpose, with
its docblock recording why. `legacy-redirect.ts` builds `nameToSlug` on it and
keeps its own Sanity imports; the analytics component imports only the pure
helper. Tests moved alongside it.
