+++ apps/web/src/lib/seo/legacy-redirect.ts
@@ -14,30 +14,14 @@
  */
 import { fetchGroq } from "@/lib/sanity/fetch-groq";
 import { SANITY_TAGS, SANITY_LIST_REVALIDATE } from "@/lib/sanity/cache-tags";
+import { slugify } from "@/lib/utils/slugify";
 
 // ─── Pure slug logic ─────────────────────────────────────────────────────────
 
 /**
- * Slugify a single string the way the legacy site keyed profile URLs:
- * lowercased, diacritics stripped, `&` → " en ", non-alphanumerics collapsed
- * to single hyphens. Mirrors the studio's `slugifyTitle` so behaviour is
- * consistent across the codebase. The general-purpose core `nameToSlug` (and
- * any other single-string slugging need, e.g. an analytics facet value)
- * builds on.
- */
-export function slugify(value: string): string {
-  return value
-    .normalize("NFKD")
-    .replace(/\p{M}+/gu, "")
-    .replace(/&/g, " en ")
-    .toLowerCase()
-    .replace(/[^a-z0-9]+/g, "-")
-    .replace(/^-+|-+$/g, "");
-}
-
-/**
  * Slugify a person's name the way the legacy site keyed profile URLs:
- * `"${firstName} ${lastName}"` through `slugify`.
+ * `"${firstName} ${lastName}"` through the shared `slugify`
+ * (`lib/utils/slugify.ts` — dependency-free on purpose; this file is not).
  */
 export function nameToSlug(firstName: string, lastName: string): string {
   return slugify(`${firstName} ${lastName}`);
